refactor: move `ui/misc` module code to `utils/aspect_ratio` module
