+++ gradia/utils/aspect_ratio.py
